vmm: move StartMicrovmError to error.rs
