+++ vmm/src/error.rs
@@ -8,11 +8,12 @@ use std::io;
 
 use super::{
     device_manager, vmm_config::boot_source::BootSourceConfigError, vmm_config::drive::DriveError,
-    vmm_config::instance_info::StartMicrovmError, vmm_config::logger::LoggerConfigError,
-    vmm_config::machine_config::VmConfigError, vmm_config::net::NetworkInterfaceError,
-    vmm_config::vsock::VsockError, vstate,
+    vmm_config::logger::LoggerConfigError, vmm_config::machine_config::VmConfigError,
+    vmm_config::net::NetworkInterfaceError, vmm_config::vsock::VsockError, vstate,
 };
 use devices::legacy::I8042DeviceError;
+use kernel::loader as kernel_loader;
+use memory_model::GuestMemoryError;
 
 /// Errors associated with the VMM internal logic. These errors cannot be generated by direct user
 /// input, but can result from bad configuration of the host (for example if Firecracker doesn't
@@ -72,6 +73,224 @@ impl std::fmt::Debug for Error {
             Serial(e) => write!(f, "Error writing to the serial console: {:?}", e),
             TimerFd(e) => write!(f, "Error creating timer fd: {}", e.to_string()),
             Vm(e) => write!(f, "Error opening VM fd: {:?}", e),
+        }
+    }
+}
+
+/// Errors associated with starting the instance.
+// TODO: add error kind to these variants because not all these errors are user or internal.
+#[derive(Debug)]
+pub enum StartMicrovmError {
+    /// This error is thrown by the minimal boot loader implementation.
+    /// It is related to a faulty memory configuration.
+    ConfigureSystem(arch::Error),
+    /// Cannot configure the VM.
+    ConfigureVm(vstate::Error),
+    /// Unable to seek the block device backing file due to invalid permissions or
+    /// the file was deleted/corrupted.
+    CreateBlockDevice(std::io::Error),
+    /// Split this at some point.
+    /// Internal errors are due to resource exhaustion.
+    /// Users errors are due to invalid permissions.
+    CreateNetDevice(devices::virtio::Error),
+    /// Failed to create a `RateLimiter` object.
+    CreateRateLimiter(std::io::Error),
+    /// Failed to create the backend for the vsock device.
+    CreateVsockBackend(devices::virtio::vsock::VsockUnixBackendError),
+    /// Failed to create the vsock device.
+    CreateVsockDevice(devices::virtio::vsock::VsockError),
+    /// The device manager was not configured.
+    DeviceManager,
+    /// Cannot read from an Event file descriptor.
+    EventFd,
+    /// Memory regions are overlapping or mmap fails.
+    GuestMemory(GuestMemoryError),
+    /// The kernel command line is invalid.
+    KernelCmdline(String),
+    /// Cannot load kernel due to invalid memory configuration or invalid kernel image.
+    KernelLoader(kernel_loader::Error),
+    #[cfg(target_arch = "x86_64")]
+    /// Cannot add devices to the Legacy I/O Bus.
+    LegacyIOBus(device_manager::legacy::Error),
+    /// Cannot load command line string.
+    LoadCommandline(kernel::cmdline::Error),
+    /// The start command was issued more than once.
+    MicroVMAlreadyRunning,
+    /// Cannot start the VM because the kernel was not configured.
+    MissingKernelConfig,
+    /// The net device configuration is missing the tap device.
+    NetDeviceNotConfigured,
+    /// Cannot open the block device backing file.
+    OpenBlockDevice(std::io::Error),
+    /// Cannot initialize a MMIO Block Device or add a device to the MMIO Bus.
+    RegisterBlockDevice(device_manager::mmio::Error),
+    /// Cannot add event to Epoll.
+    RegisterEvent,
+    /// Cannot add a device to the MMIO Bus.
+    RegisterMMIODevice(device_manager::mmio::Error),
+    /// Cannot initialize a MMIO Network Device or add a device to the MMIO Bus.
+    RegisterNetDevice(device_manager::mmio::Error),
+    /// Cannot initialize a MMIO Vsock Device or add a device to the MMIO Bus.
+    RegisterVsockDevice(device_manager::mmio::Error),
+    /// Cannot build seccomp filters.
+    SeccompFilters(seccomp::Error),
+    /// Cannot create a new vCPU file descriptor.
+    Vcpu(vstate::Error),
+    /// vCPU configuration failed.
+    VcpuConfigure(vstate::Error),
+    /// vCPUs were not configured.
+    VcpusNotConfigured,
+    /// Cannot spawn a new vCPU thread.
+    VcpuSpawn(std::io::Error),
+    /// Cannot set mode for terminal.
+    StdinHandle(std::io::Error),
+}
+
+/// It's convenient to automatically convert `kernel::cmdline::Error`s
+/// to `StartMicrovmError`s.
+impl std::convert::From<kernel::cmdline::Error> for StartMicrovmError {
+    fn from(e: kernel::cmdline::Error) -> StartMicrovmError {
+        StartMicrovmError::KernelCmdline(e.to_string())
+    }
+}
+
+impl Display for StartMicrovmError {
+    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
+        use self::StartMicrovmError::*;
+        match *self {
+            ConfigureSystem(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Faulty memory configuration. {}", err_msg)
+            }
+            ConfigureVm(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot configure virtual machine. {}", err_msg)
+            }
+            CreateBlockDevice(ref err) => write!(
+                f,
+                "Unable to seek the block device backing file due to invalid permissions or \
+                 the file was deleted/corrupted. Error number: {}",
+                err
+            ),
+            CreateRateLimiter(ref err) => write!(f, "Cannot create RateLimiter: {}", err),
+            CreateVsockBackend(ref err) => {
+                write!(f, "Cannot create backend for vsock device: {:?}", err)
+            }
+            CreateVsockDevice(ref err) => write!(f, "Cannot create vsock device: {:?}", err),
+            CreateNetDevice(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot create network device. {}", err_msg)
+            }
+            DeviceManager => write!(f, "The device manager was not configured."),
+            EventFd => write!(f, "Cannot read from an Event file descriptor."),
+            GuestMemory(ref err) => {
+                // Remove imbricated quotes from error message.
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+                write!(f, "Invalid Memory Configuration: {}", err_msg)
+            }
+            KernelCmdline(ref err) => write!(f, "Invalid kernel command line: {}", err),
+            KernelLoader(ref err) => {
+                let mut err_msg = format!("{}", err);
+                err_msg = err_msg.replace("\"", "");
+                write!(
+                    f,
+                    "Cannot load kernel due to invalid memory configuration or invalid kernel \
+                     image. {}",
+                    err_msg
+                )
+            }
+            #[cfg(target_arch = "x86_64")]
+            LegacyIOBus(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot add devices to the legacy I/O Bus. {}", err_msg)
+            }
+            LoadCommandline(ref err) => {
+                let mut err_msg = format!("{}", err);
+                err_msg = err_msg.replace("\"", "");
+                write!(f, "Cannot load command line string. {}", err_msg)
+            }
+            MicroVMAlreadyRunning => write!(f, "Microvm already running."),
+            MissingKernelConfig => write!(f, "Cannot start microvm without kernel configuration."),
+            NetDeviceNotConfigured => {
+                write!(f, "The net device configuration is missing the tap device.")
+            }
+            OpenBlockDevice(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot open the block device backing file. {}", err_msg)
+            }
+            RegisterBlockDevice(ref err) => {
+                let mut err_msg = format!("{}", err);
+                err_msg = err_msg.replace("\"", "");
+                write!(
+                    f,
+                    "Cannot initialize a MMIO Block Device or add a device to the MMIO Bus. {}",
+                    err_msg
+                )
+            }
+            RegisterEvent => write!(f, "Cannot add event to Epoll."),
+            RegisterMMIODevice(ref err) => {
+                let mut err_msg = format!("{}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot add a device to the MMIO Bus. {}", err_msg)
+            }
+            RegisterNetDevice(ref err) => {
+                let mut err_msg = format!("{}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(
+                    f,
+                    "Cannot initialize a MMIO Network Device or add a device to the MMIO Bus. {}",
+                    err_msg
+                )
+            }
+            RegisterVsockDevice(ref err) => {
+                let mut err_msg = format!("{}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(
+                    f,
+                    "Cannot initialize a MMIO Vsock Device or add a device to the MMIO Bus. {}",
+                    err_msg
+                )
+            }
+            SeccompFilters(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot build seccomp filters. {}", err_msg)
+            }
+            Vcpu(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot create a new vCPU. {}", err_msg)
+            }
+            VcpuConfigure(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "vCPU configuration failed. {}", err_msg)
+            }
+            VcpusNotConfigured => write!(f, "vCPUs were not configured."),
+            VcpuSpawn(ref err) => {
+                let mut err_msg = format!("{:?}", err);
+                err_msg = err_msg.replace("\"", "");
+
+                write!(f, "Cannot spawn vCPU thread. {}", err_msg)
+            }
+            StdinHandle(ref err) => write!(f, "Failed to set mode for terminal: {}", err),
         }
     }
 }
